Refactor NoteDataBase to utilize dependency injection
