+++ solutions/devsprint-higor-silva-2/app/src/main/java/com/devpass/mynotes/data/NoteDataBase.kt
@@ -1,29 +1,12 @@
 package com.devpass.mynotes.data
 
-import android.content.Context
 import androidx.room.Database
-import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.devpass.mynotes.domain.model.Note
 
 @Database(entities = [Note::class], version = 1, exportSchema = false)
 abstract class NoteDataBase: RoomDatabase() {
 
-    companion object {
-
-        @Volatile
-        private var INSTANCE: NoteDataBase? = null
-
-        fun getDataBase(context: Context): NoteDataBase {
-            return INSTANCE ?: synchronized(this) {
-                val instance = Room.databaseBuilder(
-                    context.applicationContext,
-                    NoteDataBase::class.java,
-                    "note_database"
-                ).build()
-                INSTANCE = instance
-                instance
-            }
-        }
-    }
+    abstract fun getNotesDao(): NotesDao
 
 }
