+++ solutions/devsprint-higor-silva-2/app/src/main/java/com/devpass/mynotes/data/NotesDao.kt
@@ -7,6 +7,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
+import com.devpass.mynotes.domain.model.Note
 
 
 @Dao
