+++ solutions/devsprint-higor-silva-2/app/src/main/java/com/devpass/mynotes/di/AppModule.kt
@@ -1,3 +1,33 @@
 package com.devpass.mynotes.di
 
-object AppModule
+import android.content.Context
+import androidx.room.Room
+import com.devpass.mynotes.data.NoteDataBase
+import com.devpass.mynotes.data.NotesDao
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@InstallIn(SingletonComponent::class)
+@Module
+object AppModule {
+
+    @Singleton
+    @Provides
+    fun provideNoteDao(noteDataBase: NoteDataBase): NotesDao {
+        return noteDataBase.getNotesDao()
+    }
+
+    @Provides
+    @Singleton
+    fun provideNoteDatabase(@ApplicationContext appContext: Context): NoteDataBase {
+        return Room.databaseBuilder(
+            appContext,
+            NoteDataBase::class.java,
+            "note_table"
+        ).build()
+    }
+}
